Included freesurfer directory path
Previously the code had assumed that the freesurfer directory was saved
inside the HIGHRES folder, but that isn't always the case so now this
code takes the location of the freesurfer directory as an argument. All
registrations are still in the REG directory at the same level as the
HIGHRES folder though.